requirements.txt
fastapi==0.121.3
google-adk==1.18.0
google-generativeai==0.8.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google_api_python_client==2.187.0
google_auth_httplib2==0.2.1
gspread==6.2.1
httplib2==0.31.0
litellm
httpx==0.28.1
pandas==2.3.3
protobuf==6.33.1
pydantic==2.12.4
python-dotenv==1.2.1
python-telegram-bot==22.5
uvicorn==0.38.0
psycopg2-binary==2.9.10
asyncpg==0.29.0
oauth2client==4.1.3
Deprecated==1.2.14
duckduckgo-search